+++ kernel/kernel.c
@@ -7,59 +7,63 @@
 #include "tls.h"
 
 extern void asm_hlt(void);
-/*
-typedef struct {
+
+#define INT_TYPE_INTERRUPT 0x8e
+#define INT_TYPE_TRAP 0x8f
+
+typedef struct __packed {
+	uint16_t offset1;        // offset bits 0..15
+	uint16_t selector;        // a code segment selector in GDT or LDT
+	uint8_t  ist;             // bits 0..2 holds Interrupt Stack Table offset, rest of bits zero.
+	uint8_t  type_attributes; // gate type, dpl, and p fields
+	uint16_t offset2;        // offset bits 16..31
+	uint32_t offset3;        // offset bits 32..63
+	uint32_t zero;            // reserved
+} IntDesc64;
+
+typedef struct __packed {
 	uint16_t limit;
-	void    *base;
-} __packed idtr_t;
+	IntDesc64 *base;
+} idtr_t;
 
-idtr_t itdr;
-
-static const char* memtypeconvert(EFI_MEMORY_TYPE memtype){
-	switch (memtype){
-	case EfiReservedMemoryType:
-		return "reserved";
-	case EfiLoaderCode:
-		return "ldr-code";
-	case EfiLoaderData:
-		return "ldr-data";
-	case EfiBootServicesCode:
-		return "BS-code ";
-	case EfiBootServicesData:
-		return "BS-data ";
-	case EfiRuntimeServicesCode:
-		return "RT-code ";
-	case EfiRuntimeServicesData:
-		return "RT-data ";
-	case EfiConventionalMemory:
-		return "conv-mem";
-	case EfiUnusableMemory:
-		return "unusable";
-	case EfiACPIReclaimMemory:
-		return "ACPIrecl";
-	case EfiACPIMemoryNVS:
-		return "ACPI-NVS";
-	case EfiMemoryMappedIO:
-		return "MMIO    ";
-	case EfiMemoryMappedIOPortSpace:
-		return "MMIO-PS ";
-	case EfiPalCode:
-		return "palcode ";
-	default:
-		return "unknown ";
-	}
-}
-*/
+idtr_t idtr __attribute__((aligned(16)));
 
 uint64_t __tls test_thread_data = 0x0102030405060708ull;
+
+extern void isr_start();
+
+static uint16_t repeat = 0;
+
+uint8_t isr_common(uint8_t intn, uint64_t* stack){
+	printf("%u %p %p %p %p %p %p\n", (uint64_t)intn, *(stack), *(stack+1), *(stack+2), *(stack+3), *(stack+4), *(stack+5));
+	if (intn == repeat){
+		for (;;);
+	}
+	else if (intn < 32){
+		repeat = intn;
+		return 1;
+	}
+	return 0;
+}
+
+__export void fill_idt(IntDesc64 *ent, uint64_t ptr, uint16_t ss, uint8_t ist, uint8_t type){
+	ent->offset1 = ptr & 0xFFFF;
+	ent->offset2 = (ptr >> 16) & 0xFFFF;
+	ent->offset3 = (uint32_t)(ptr >> 32);
+
+	ent->selector = ss;
+	ent->ist = ist;
+	ent->type_attributes = type;
+	ent->zero = 0;
+}
 
 void kmain(){
 	init_video(boot_info->vram, boot_info->width, boot_info->height, boot_info->ppl, boot_info->format);
 	fill_rect(0, 0, 0, boot_info->width-1, boot_info->height-1);
 	/*
-	itdr.limit = 4095;
-	itdr.base = (void*)boot_info - 4096;
-	asm volatile ("lidt %0" : : "m"(itdr));
+	idtr.limit = 4095;
+	idtr.base = (void*)boot_info - 4096;
+	asm volatile ("lidt %0" : : "m"(idtr));
 	*/
 	set_color(COLOR_WHITE, COLOR_BLACK);
 
@@ -97,7 +101,7 @@ void kmain(){
 	// for(size_t i=0; i < e820_len; i++){
 	// 	printf("%p | %p | %u    | %u\n" , e820[i].base, e820[i].lenght, e820[i].type, e820[i].attr);
 	// }
-
+/*
 	set_color(COLOR_WHITE, COLOR_BLACK);
 	printf("gorl: \n");
 	for (size_t i = 0; i < gorl.lenght; i++){
@@ -126,9 +130,22 @@ void kmain(){
 	uint64_t gsbase;
 	rdmsr(IA32_GS_BASE, gsbase);
 	printf("GS BASE: %p\n", gsbase);
-
 	printf("TLS TEST: %p\n", *tls_ptr(test_thread_data));
 	//printf("TLS TEST: %p\n", test_global_data);
+*/
+	idtr.base = allocate_pages(1);
+	idtr.limit = 4095;
+	for (uint16_t i = 0; i < 256; i++){
+		fill_idt(&idtr.base[i], (uint64_t)isr_start + i*7, 0x8, 0, INT_TYPE_INTERRUPT);
+	}
+
+	asm volatile ("lidt %0 \n\t" :: "m"(idtr) : "memory");
+	printf("IDT loaded! \n");
+
+	asm volatile ("sti \n\t" ::: "memory");
+	printf("Interrupts! \n");
+
+	asm volatile ("int $255 \n\t" ::: "memory");
 
 	//for(;;);
 	asm_hlt();